requirements.txt
langchain-google-genai
langchain-core
langchain-community
python-dotenv
pypdf
Pillow
chainlit
